pyproject.toml
[project]
name = "mcp-map-server"
version = "0.1.0"
description = ""
authors = [
    {name = "soRan0219",email = "124494718+soRan0219@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastmcp (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "mcp_map_server", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
